refactor: mise à jour des types et des imports pour la gestion des écoles; amélioration de la logique de navigation et de création d'école
